requirements.txt
asgiref==3.8.1
attrs==25.1.0
Automat==25.4.16
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.4.0
outcome==1.3.0.post0
packaging==25.0
parsel==1.10.0
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
PyJWT==2.9.0
pyOpenSSL==25.0.0
PyPDF2==3.0.1
PySocks==1.7.1
queuelib==1.8.0
requests==2.32.3
requests-file==2.1.0
scrapingbee==2.0.1
Scrapy==2.12.0
selenium==4.29.0
service-identity==24.2.0
setuptools==80.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tldextract==5.3.0
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
w3lib==2.3.1
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
zope.interface==7.2
